feat: add repository implementations and ViewModels for MVI
- Implement IngredientRepositoryImpl, RecipeRepositoryImpl, AdditionalVariableCostRepositoryImpl, FixedCostRepositoryImpl, SettingsRepositoryImpl
- Add CalculateRecipeCostUseCase with recursive ancestor chain for recipe costing
- Create MVI ViewModels: IngredientesViewModel, RecetasViewModel, CostosFijosViewModel, DashboardViewModel
- Fix SQLDelight type mappings (Long vs Int)
- Fix Settings column naming (value_ suffix)
- Update RecipeIngredient.yieldPizzas to Long

+++ shared/src/commonMain/kotlin/com/recipecostcalculator/domain/model/RecipeIngredient.kt
@@ -6,5 +6,5 @@ data class RecipeIngredient(
     val ingredientId: Long,
     val ingredient: Ingredient? = null,
     val usagePerPizza: Double?,
-    val yieldPizzas: Int?
+    val yieldPizzas: Long?
 )
